[refactor]: deleted models.py, moved its contents to enums.py. Added missing comments describing functions and classes
