+++ src/odt/enum/namespaces.py
@@ -1,0 +1,7 @@
+class NameSpaces:
+    """Class with document namespace odt"""
+    STYLE = '{urn:oasis:names:tc:opendocument:xmlns:style:1.0}'
+    TEXT = '{urn:oasis:names:tc:opendocument:xmlns:text:1.0}'
+    FO = '{urn:oasis:names:tc:opendocument:xmlns:xsl-fo-compatible:1.0}'
+    OFFICE = '{urn:oasis:names:tc:opendocument:xmlns:office:1.0}'
+    XML = '{http://www.w3.org/XML/1998/namespace}'
